requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.6.1
sqlalchemy==2.0.25
python-multipart==0.0.6
groq==0.4.2
whisper==1.1.10
gTTS==2.5.1
numpy
torch==2.2.0
soundfile==0.12.1
librosa==0.10.1
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aiofiles==23.2.1
nltk==3.8.1
websockets==12.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0 
groq